Move getDependencyVersion to utils/dependencies.ts
